dev-requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile '.\dev-requirements.in'
#
appdirs==1.4.3            # via black, virtualenv
astroid==2.3.3            # via pylint
atomicwrites==1.3.0       # via pytest
attrs==19.3.0             # via black, flake8-bugbear, hypothesis, pytest
azure-functions==1.2.0    # via -r .\__app__/requirements.txt, -r .\dev-requirements.in
bandit==1.6.2             # via flake8-bandit
black==19.10b0            # via -r .\dev-requirements.in
certifi==2020.4.5.1       # via pipenv, requests
chardet==3.0.4            # via requests
click==7.1.1              # via black, pip-tools, safety
colorama==0.4.3           # via bandit, pylint, pytest
coverage==5.1             # via pytest-cov
distlib==0.3.0            # via virtualenv
dparse==0.5.0             # via safety
entrypoints==0.3          # via flake8
filelock==3.0.12          # via virtualenv
flake8-bandit==2.1.2      # via -r .\dev-requirements.in
flake8-bugbear==20.1.4    # via -r .\dev-requirements.in
flake8-deprecated==1.3    # via -r .\dev-requirements.in
flake8-docstrings==1.5.0  # via -r .\dev-requirements.in
flake8-executable==2.0.3  # via -r .\dev-requirements.in
flake8-polyfill==1.0.2    # via flake8-bandit
flake8==3.7.9             # via -r .\dev-requirements.in, flake8-bandit, flake8-bugbear, flake8-deprecated, flake8-docstrings, flake8-executable, flake8-polyfill
gitdb==4.0.4              # via gitpython
gitpython==3.1.1          # via bandit
hypothesis==5.10.4        # via -r .\dev-requirements.in
idna==2.9                 # via requests
importlib-resources==1.4.0  # via -r .\dev-requirements.in
isort==4.3.21             # via pylint
lazy-object-proxy==1.4.3  # via astroid
mccabe==0.6.1             # via flake8, pylint
more-itertools==8.2.0     # via pytest
mypy-extensions==0.4.3    # via mypy
mypy==0.770               # via -r .\dev-requirements.in
packaging==20.3           # via dparse, pytest, safety
pathspec==0.8.0           # via black
pbr==5.4.5                # via stevedore
pip-tools==5.0.0          # via -r .\dev-requirements.in
pipenv==2018.11.26        # via dparse
pluggy==0.13.1            # via pytest
py==1.8.1                 # via pytest
pycodestyle==2.5.0        # via flake8, flake8-bandit
pydocstyle==5.0.2         # via flake8-docstrings
pyflakes==2.1.1           # via flake8
pylint==2.4.4             # via -r .\dev-requirements.in
pyparsing==2.4.7          # via packaging
pytest-asyncio==0.11.0    # via -r .\dev-requirements.in
pytest-cov==2.8.1         # via -r .\dev-requirements.in
pytest==5.4.1             # via -r .\dev-requirements.in, pytest-asyncio, pytest-cov
pyyaml==5.3.1             # via bandit, dparse
regex==2020.4.4           # via black
requests==2.23.0          # via safety
rope==0.16.0              # via -r .\dev-requirements.in
safety==1.8.7             # via -r .\dev-requirements.in
six==1.14.0               # via astroid, bandit, packaging, pip-tools, stevedore, virtualenv
smmap==3.0.2              # via gitdb
snowballstemmer==2.0.0    # via pydocstyle
sortedcontainers==2.1.0   # via hypothesis
stevedore==1.32.0         # via bandit
toml==0.10.0              # via black, dparse
typed-ast==1.4.1          # via black, mypy
typing-extensions==3.7.4.2  # via mypy
urllib3==1.25.9           # via requests
virtualenv-clone==0.5.4   # via pipenv
virtualenv==20.0.18       # via pipenv
wcwidth==0.1.9            # via pytest
wrapt==1.11.2             # via astroid

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
